Align deck creation UI props with main
